requirements.txt
aiohttp==3.11.10
aiosqlite==0.20.0
asyncpg==0.30.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
SQLAlchemy==2.0.44
